Add 2.0 second sleep on 503 error
The docs for this project encourage users to do things like

ia search 'subject:"market street" collection:prelinger' | parallel -j40 'ia metadata {}'

If the webserver starts throwing 503 errors, then we need to reduce
the amount of work we're doing. Since the 'ia metadata' command in the
above command line is only downloading one thing, the most reasonable
way to slow things down is to add a sleep() if a 503 error is seen.