Cast all headers to string
Some headers were being passed to requests with integer values, which
would cause requests to raise a InvalidHeader error (at line 799 of
requests/utils.py).  This casts all headers to string, which solves the
issue.  I tested this with an upload to the Internet Archive and it
avoided the error for me - was not able to get uploading to work
otherwise.