Fixed bug in get_s3_xml_text().
``internetarchive.utils.get_s3_xml_text()`` is used to parse errors
returned by S3 in XML. Sometimes there is no XML in the response. Most
of the time this is due to 5xx errors. Either way, we want to always
return the HTTPError, even if the XML parsing fails. This changes fixes
that.