Add a metadata attribute to File.
This enables direct access to the file metadata without
having to extract the relevant metadata keys from the
object's attributes.

It is the most maintainable choice in case of future
additions like new non-metadata attributes.

Since the Internet Archive always creates some attributes
on the fly (e. g. the checksums and file size), those do not
have to be filtered out when copying the metadata, because they
will be overwritten.