Specify the maximum length of identifiers.
The other Internet Archive documentation specifies the maximum length
as 100 characters, but actually it is 101 characters as you can test
by creating such an item and trying to create one which is greater
than 101 characters.

This is probably an off-by-one error in the Internet Archive program
which checks identifiers.

See: https://archive.org/services/docs/api/metadata-schema/index.html