Replace two deprecated names with their supported names.
The keyword parameter `method_whitelist` and the attribute
`DEFAULT_METHOD_WHITELIST` of `Retry` were renamed to
`allowed_methods` and `DEFAULT_ALLOWED_METHODS`.
The tests suit is showing warnings due to their deprecation.

The old names will be removed in version 2.0 of urllib3.
It is noteworthy that the Python 2 support will also be dropped
when version 2.0 is reached.