Expand a period to the contents of the current directory.
The previous behavior when specifying a period as a wildcard was faulty.
Now a period is properly expanded to the contents of the current directory.

ia_upload.py seemed to be the most sensible location for this fix, since it
is the closest to receiving the path and so that the behavior of the library
does not change.

Closes #220, fixes #535.