Merge the period expanded directory and regular pathes.
This does not make too much sense, yet it is the least surprising variant.

The period is a shortcut to upload the content of the current
directory (and thereby its sub-directories). If someone wanted to
upload the content of multiple directories to the same item it would
be strange if they ran the command in one of multiple equivalent
directories and used the dot for that one directory when they had to
supply the pathes of all the others anyways.
Specifying an additional file in the directory which is already
covered by the period expansion obviously does not make any sense
either.

But now people can do those things and get exactly what they would expect.