Skip early when downloading existing file
#614 moved skip checks to until after response headers have been
received, which drastically slows down the download process if
the file already exists or the file has an equal checksum. Since
the file name and checksum are already known prior to download,
these checks should remain at the start to avoid having to make
a request which would eventually be discarded anyway.